requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
joblib
ta          # <--- THIS IS THE CORRECT ONE NOW
ccxt
shap
optuna